+++ packages/ui_flutter/lib/src/components/select/select.dart
@@ -1,4 +1,10 @@
 /// Tinyrack select component public API.
 library;
 
-export 'select_widget.dart' show TRSelect, TRSelectFormField, TRSelectItem;
+export 'select_widget.dart'
+    show
+        TRSelect,
+        TRSelectFilter,
+        TRSelectFormField,
+        TRSelectItem,
+        TRSelectSurface;